+++ src/main/java/com/webtoon/content/repository/ContentRepository.java
@@ -11,5 +11,5 @@ public interface ContentRepository {
 
     Content getById(Long id);
 
-    Optional<Content> findByCartoonIdAndEpisode(Long cartoonId, Long episode);
+    Optional<Content> findByCartoonIdAndEpisode(Long cartoonId, int episode);
 }
